+++ app/modules/services/FileSuggestionService/Module.swift
@@ -2,23 +2,15 @@ Target.module(
   name: "FileSuggestionService",
   dependencies: [
     .product(name: "Ifrit", package: "Ifrit"),
-    .product(name: "XcodeProj", package: "XcodeProj"),
     "AppFoundation",
     "ConcurrencyFoundation",
     "DependencyFoundation",
     "FileSuggestionServiceInterface",
-    "FoundationInterfaces",
-    "ShellServiceInterface",
     "ThreadSafe",
     "XcodeObserverServiceInterface",
   ],
   testDependencies: [
     "ConcurrencyFoundation",
-    "FoundationInterfaces",
-    "ShellServiceInterface",
     "SwiftTesting",
     "XcodeObserverServiceInterface",
-  ],
-  testResources: [
-    .copy("resources/"),
   ])
