+++ app/modules/services/XcodeObserverService/Module.swift
@@ -1,6 +1,7 @@
 Target.module(
   name: "XcodeObserverService",
   dependencies: [
+    .product(name: "XcodeProj", package: "XcodeProj"),
     "AccessibilityFoundation",
     "AppFoundation",
     "ConcurrencyFoundation",
@@ -13,4 +14,16 @@ Target.module(
     "ThreadSafe",
     "XcodeObserverServiceInterface",
   ],
-  testDependencies: [])
+  testDependencies: [
+    "ConcurrencyFoundation",
+    "FoundationInterfaces",
+    "LoggingServiceInterface",
+    "PermissionsServiceInterface",
+    "SettingsServiceInterface",
+    "ShellServiceInterface",
+    "SwiftTesting",
+    "XcodeObserverServiceInterface",
+  ],
+  testResources: [
+    .copy("resources/"),
+  ])
