+++ app/modules/services/XcodeObserverService/Tests/resources/SPM/.gitignore
