+++ app/modules/services/XcodeObserverService/Tests/resources/SPM/Package.swift
