+++ app/modules/services/XcodeObserverService/Tests/resources/TestXcodeProjParsing/TestXcodeProjParsing.xcodeproj/project.xcworkspace/contents.xcworkspacedata
