+++ app/modules/services/XcodeObserverService/Tests/resources/TestXcodeProjParsing/TestXcodeProjParsing/Assets.xcassets/AccentColor.colorset/Contents.json
