+++ app/modules/services/XcodeObserverService/Tests/resources/TestXcodeProjParsing/TestXcodeProjParsing/Assets.xcassets/AppIcon.appiconset/Contents.json
