+++ app/modules/services/XcodeObserverService/Tests/resources/TestXcodeProjParsing/TestXcodeProjParsing/ContentView.swift
