+++ app/modules/services/XcodeObserverService/Tests/resources/TestXcodeProjParsing/TestXcodeProjParsing/Info.plist
