+++ app/modules/services/XcodeObserverService/Tests/resources/TestXcodeProjParsing/TestXcodeProjParsing/Preview Content/Preview Assets.xcassets/Contents.json
