+++ app/modules/services/XcodeObserverService/Tests/resources/TestXcodeProjParsing/TestXcodeProjParsing/TestXcodeProjParsingApp.swift
